Chore: Move stepperV2 to tracer
