+++ src/tracer/nodes/Statement/FunctionDeclaration.ts
