+++ src/tracer/nodes/Statement/VariableDeclaration.ts
